Add one more null check in Pawn_Jobtracker
Not all animals have a rest meter, for example animals from Android Tiers.
I added this one more null check so that animals with no rest bars don't throw NRE while resting.

This also revealed another non-bug issue: There are no checks here for if animal is resting because it is sick (with diseases like flu, since they don't affect SummaryHealthPercent), or because it has a surgery assigned. The animal would be considered available when IMO it shouldn't. This issue don't immediately break the mod though so I didn't bother to propose a fix.